requirements.txt
paho-mqtt>=1.6.1
pyserial>=3.5
pyyaml>=6.0
asyncio-mqtt>=0.13.0
aiofiles>=23.1.0
click>=8.1.0
colorlog>=6.7.0 